requirements.txt
asgiref==3.11.0
asttokens==3.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
coreapi==2.3.3
coreschema==0.0.4
decorator==5.2.1
dj-database-url==3.0.1
Django==6.0
django-debug-toolbar==6.1.0
django-filter==25.2
django-health-check==3.20.0
django-querycount==0.8.3
django-unfold==0.72.0
djangorestframework==3.16.1
drf-redesign==0.6.0
drf-yasg==1.21.11
executing==2.2.1
factory_boy==3.3.3
Faker==38.2.0
idna==3.11
inflection==0.5.1
ipdb==0.13.13
ipython==9.8.0
ipython_pygments_lexers==1.1.1
itypes==1.2.0
jedi==0.19.2
Jinja2==3.1.6
Markdown==3.10
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
openapi-codec==1.3.2
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
simplejson==3.20.2
sqlparse==0.5.4
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.1
wcwidth==0.2.14
